Split out middleware from services
