+++ middleware/auth.js
@@ -1,10 +1,10 @@
 'use strict';
 const config = require('config');
 const cookieParser = require('cookie-parser');
-const OIDCStrategy = require('passport-azure-ad').OIDCStrategy;
-const passport = require('passport');
 const session = require('express-session');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
+const passport = require('passport');
+const OIDCStrategy = require('passport-azure-ad').OIDCStrategy;
 
 const models = require('../models');
 const userService = require('../services/user');
@@ -35,14 +35,7 @@ function authMiddlewareLogin(req, res, next) {
     })(req, res, next);
 }
 
-/**
- * Set up global auth middleware
- * - Initialise session
- * - Configure passport strategy
- */
 function globalMiddleware() {
-    const secret = process.env.SESSION_SECRET;
-
     const store = new SequelizeStore({
         db: models.sequelize
     });
@@ -50,6 +43,7 @@ function globalMiddleware() {
     // create sessions table
     store.sync();
 
+    const secret = process.env.SESSION_SECRET;
     const sessionConfig = {
         name: 'blf-applications',
         secret: secret,
@@ -58,52 +52,50 @@ function globalMiddleware() {
         store: store
     };
 
-    const oidcConfig = {
-        identityMetadata: config.get('auth.identityMetadata'),
-        clientID: config.get('auth.clientID'),
-        responseType: config.get('auth.responseType'),
-        responseMode: config.get('auth.responseMode'),
-        redirectUrl: config.get('auth.redirectUrl'),
-        allowHttpForRedirectUrl: config.get('auth.allowHttpForRedirectUrl'),
-        clientSecret: config.get('auth.clientSecret'),
-        validateIssuer: config.get('auth.validateIssuer'),
-        isB2C: config.get('auth.isB2C'),
-        issuer: config.get('auth.issuer'),
-        passReqToCallback: config.get('auth.passReqToCallback'),
-        scope: config.get('auth.scope'),
-        loggingLevel: config.get('auth.loggingLevel'),
-        nonceLifetime: config.get('auth.nonceLifetime'),
-        nonceMaxAmount: config.get('auth.nonceMaxAmount'),
-        useCookieInsteadOfSession: config.get('auth.useCookieInsteadOfSession'),
-        cookieEncryptionKeys: config.get('auth.cookieEncryptionKeys'),
-        clockSkew: config.get('auth.clockSkew')
-    };
-
     passport.use(
-        new OIDCStrategy(oidcConfig, (iss, sub, profile, accessToken, refreshToken, done) => {
-            if (!profile.oid) {
-                return done(new Error('No oid found'), null);
-            }
-            // asynchronous verification, for effect...
-            process.nextTick(() => {
-                userService.findUser(profile.oid, (err, user) => {
-                    if (err) {
-                        return done(err);
-                    }
-                    if (!user) {
-                        userService
-                            .createUser(profile)
-                            .then(() => {
+        new OIDCStrategy(
+            {
+                identityMetadata: config.get('auth.identityMetadata'),
+                clientID: config.get('auth.clientID'),
+                responseType: config.get('auth.responseType'),
+                responseMode: config.get('auth.responseMode'),
+                redirectUrl: config.get('auth.redirectUrl'),
+                allowHttpForRedirectUrl: config.get('auth.allowHttpForRedirectUrl'),
+                clientSecret: config.get('auth.clientSecret'),
+                validateIssuer: config.get('auth.validateIssuer'),
+                isB2C: config.get('auth.isB2C'),
+                issuer: config.get('auth.issuer'),
+                passReqToCallback: config.get('auth.passReqToCallback'),
+                scope: config.get('auth.scope'),
+                loggingLevel: config.get('auth.loggingLevel'),
+                nonceLifetime: config.get('auth.nonceLifetime'),
+                nonceMaxAmount: config.get('auth.nonceMaxAmount'),
+                useCookieInsteadOfSession: config.get('auth.useCookieInsteadOfSession'),
+                cookieEncryptionKeys: config.get('auth.cookieEncryptionKeys'),
+                clockSkew: config.get('auth.clockSkew')
+            },
+            (iss, sub, profile, accessToken, refreshToken, done) => {
+                if (!profile.oid) {
+                    return done(new Error('No oid found'), null);
+                }
+                // asynchronous verification, for effect...
+                process.nextTick(() => {
+                    userService.findUser(profile.oid, (err, user) => {
+                        if (err) {
+                            return done(err);
+                        }
+                        if (!user) {
+                            userService.createUser(profile).then(() => {
                                 return done(null, profile);
-                            })
-                            .catch(() => {
+                            }).catch(() => {
                                 return done(null, user);
                             });
-                    }
-                    return done(null, user);
+                        }
+                        return done(null, user);
+                    });
                 });
-            });
-        })
+            }
+        )
     );
 
     passport.serializeUser((user, done) => {
@@ -116,7 +108,7 @@ function globalMiddleware() {
         });
     });
 
-    return [passport.initialize(), passport.session(), cookieParser(secret), session(sessionConfig)];
+    return [cookieParser(secret), session(sessionConfig), passport.initialize(), passport.session()];
 }
 
 module.exports = {
