--- services/bodyParser.js
@@ -1,4 +1,0 @@
-'use strict';
-const bodyParser = require('body-parser');
-
-module.exports = [bodyParser.json(), bodyParser.urlencoded({ extended: false })];
